[CAN-33] chore: NavProfile -> NavUserProfile로 이름 변경
